% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcualate_growth_rate_and_expression_Biotek.R
\name{calcualate_growth_rate_and_expression_Biotek}
\alias{calcualate_growth_rate_and_expression_Biotek}
\title{calcualate_growth_rate_and_expression_Biotek
#'
A function use output of plate reader to caluclate growth rate, if have fluorescence data, it can also calculate fluorescence expression level.}
\usage{
calcualate_growth_rate_and_expression_Biotek(
  inputfile,
  skip_num,
  plate,
  read_times_per_hour,
  total_read_times,
  blank_wells,
  X_hours_SW,
  medimum1 = "carbon",
  medimum2 = "nitrogen",
  outputdir,
  outputfile,
  outputfile2
)
}
\arguments{
\item{inputfile}{inputfile="./dat-001-001-sd-p121.123.3AT_SCD-H.xlsx";}

\item{skip_num}{skip_num=40, how many lines before the real data}

\item{plate}{plate=96, means 96 well plate, plate=384 means 384 well plate.}

\item{read_times_per_hour}{read_times_per_hour=6, how many points in one hour.}

\item{total_read_times}{total_read_times=217, total points in this experiment.}

\item{blank_wells}{blank_wells=c(1:12); the wells only with medimum as control. For example, A01:A12 (first row) is 1:12 and A01:H01 (first column) is c(1, 13, 25, 37, 49, 61, 73, 85)}

\item{X_hours_SW}{X_hours_SW=2 The time length (hours) of sliding window, need to be less than the total time. The total time points used is X_hours_SW*read_times_per_hour. Suggest to make X_hours_SW\*read_times_per_hour >= 10}

\item{medimum1}{Defaults to medimum1="carbon"; used to label samples.}

\item{medimum2}{Defaults to medimum2="nitrogen"; used to label samples.}

\item{outputdir}{outputdir="./20200303output/";  the folder you want to output}

\item{outputfile}{outputfile= "20200303saad.xlsx"; xlsx format, the growth rate result file}

\item{outputfile2}{outputfile2 = "./20190731test96fi_2.xlsx"; xlsx format, the fluoresence result file, if no, set like this.}
}
\description{
calcualate_growth_rate_and_expression_Biotek
#'
A function use output of plate reader to caluclate growth rate, if have fluorescence data, it can also calculate fluorescence expression level.
}
\examples{
calcualate_growth_rate(inputfile,skip_num, plate,read_times_per_hour,total_read_times, blank_wells,X_hours_SW,medimum1="carbon", medimum2="nitrogen",outputdir,outputfile,outputfile2)
}
\keyword{growth_rate}
